AI-Agents-CrewAI-zip_-KeIq/requirements.txt
crewai==0.22.5
crewai[tools]
langchain_openai
langchain_community
python-dotenv
langchain_groq
#pip install -U crewai crewai-tools

